refactor: simplify code and move functions out of STEP
Moved the code displaying text on the screen into a dedicated function,
thus cleaning up the main loop of the program
Moved functions, previously residing within the STEP loop, out and
up in DRAW(). This allows the screen and particles to
show regardless of whether the user steps through the code. The idea
is to eventually only have the robot, agent and particle updates
within the STEP loop